requirements.txt
Flask
python-dotenv
line-bot-sdk==3.11.0
requests
pymongo[srv]
chromadb==0.5.2